refactor : graphpage - nodeslidebar
